Job Cloud/task_a/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.18.2
google-cloud-pubsub==2.21.6
